[FAILS] Update e2e call tests to regular call tests
e2e tests still fail, injections code generation not implemented yet
