--- lang/lang.test/src/test/java/mb/pie/lang/test/call/contextParams/T1.java
@@ -1,12 +1,0 @@
-package mb.pie.lang.test.call.contextParams;
-
-import javax.inject.Inject;
-
-public class T1 {
-    @Inject
-    public T1() {}
-
-    public int get() {
-        return 1;
-    }
-}
