--- lang/lang.test/src/test/java/mb/pie/lang/test/call/contextParams/T2.java
@@ -1,12 +1,0 @@
-package mb.pie.lang.test.call.contextParams;
-
-import javax.inject.Inject;
-
-public class T2 {
-    @Inject
-    public T2() {}
-
-    public int get() {
-        return 2;
-    }
-}
