--- lang/lang.test/src/test/java/mb/pie/lang/test/call/contextParams/empty/EmptyTest.java
@@ -1,13 +1,0 @@
-package mb.pie.lang.test.call.contextParams.empty;
-
-import mb.pie.api.ExecException;
-import mb.pie.lang.test.call.contextParams.empty.DaggeremptyComponent;
-import org.junit.jupiter.api.Test;
-
-import static mb.pie.lang.test.util.SimpleChecker.assertTaskOutputEquals;
-
-class EmptyTest {
-    @Test void test() throws ExecException {
-        assertTaskOutputEquals(DaggeremptyComponent.class, 0);
-    }
-}
