--- lang/lang.test/src/test/java/mb/pie/lang/test/call/contextParams/three/three.pie
@@ -1,12 +1,0 @@
-module mb:pie:lang:test:call:contextParams:three
-
-data T1 = foreign java mb.pie.lang.test.call.contextParams.T1 {
-  func get() -> int
-}
-data T2 = foreign java mb.pie.lang.test.call.contextParams.T2 {
-  func get() -> int
-}
-
-func main_three() -> int = helper_three()
-func helper_three()[p1: T1, p2: T2, p3: T1] -> int =
-  p1.get() + (p2.get()) + (p3.get())
