--- lang/lang.test/src/test/java/mb/pie/lang/test/call/contextParams/two/two.pie
@@ -1,11 +1,0 @@
-module mb:pie:lang:test:call:contextParams:two
-
-data T1 = foreign java mb.pie.lang.test.call.contextParams.T1 {
-  func get() -> int
-}
-data T2 = foreign java mb.pie.lang.test.call.contextParams.T2 {
-  func get() -> int
-}
-
-func main_two() -> int = helper_two()
-func helper_two()[p1: T1, p2: T2] -> int = p1.get() + (p2.get())
