+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/nullary/NullaryTest.java
@@ -1,0 +1,13 @@
+package mb.pie.lang.test.call.pieFunc.nullary;
+
+import mb.pie.api.ExecException;
+import mb.pie.lang.test.call.pieFunc.nullary.DaggernullaryComponent;
+import org.junit.jupiter.api.Test;
+
+import static mb.pie.lang.test.util.SimpleChecker.assertTaskOutputEquals;
+
+class NullaryTest {
+    @Test void test() throws ExecException {
+        assertTaskOutputEquals(DaggernullaryComponent.class, 0);
+    }
+}
