+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/nullary/nullary.pie
@@ -1,0 +1,4 @@
+module mb:pie:lang:test:call:pieFunc:nullary
+
+func main_nullary() -> int = helper_nullary()
+func helper_nullary() -> int = 0
