+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/one/PieTestModule.java
@@ -1,10 +1,12 @@
-package mb.pie.lang.test.call.contextParams.empty;
+package mb.pie.lang.test.call.pieFunc.one;
 
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.ElementsIntoSet;
 import mb.pie.api.TaskDef;
-import mb.pie.lang.test.call.contextParams.empty.main_empty;
+import mb.pie.lang.test.call.pieFunc.empty.helper_empty;
+import mb.pie.lang.test.call.pieFunc.empty.main_empty;
+import mb.pie.lang.test.call.pieFunc.one.main_one;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -13,8 +15,8 @@ import java.util.Set;
 abstract class PieTestModule {
     @Provides @mb.pie.dagger.PieScope @ElementsIntoSet
     public static Set<TaskDef<?, ?>> provideTaskDefs(
-        main_empty main,
-        helper_empty helper
+        main_one main,
+        helper_one helper
     ) {
         final HashSet<TaskDef<?, ?>> taskDefs = new HashSet<>(2, 1);
         taskDefs.add(main);
