+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/three/PieTestModule.java
@@ -1,12 +1,12 @@
-package mb.pie.lang.test.call.contextParams.one;
+package mb.pie.lang.test.call.pieFunc.three;
 
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.ElementsIntoSet;
 import mb.pie.api.TaskDef;
-import mb.pie.lang.test.call.contextParams.empty.helper_empty;
-import mb.pie.lang.test.call.contextParams.empty.main_empty;
-import mb.pie.lang.test.call.contextParams.one.main_one;
+import mb.pie.lang.test.call.pieFunc.empty.helper_empty;
+import mb.pie.lang.test.call.pieFunc.empty.main_empty;
+import mb.pie.lang.test.call.pieFunc.three.main_three;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -15,8 +15,8 @@ import java.util.Set;
 abstract class PieTestModule {
     @Provides @mb.pie.dagger.PieScope @ElementsIntoSet
     public static Set<TaskDef<?, ?>> provideTaskDefs(
-        main_one main,
-        helper_one helper
+        main_three main,
+        helper_three helper
     ) {
         final HashSet<TaskDef<?, ?>> taskDefs = new HashSet<>(2, 1);
         taskDefs.add(main);
