+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/three/ThreeTest.java
@@ -1,13 +1,13 @@
-package mb.pie.lang.test.call.contextParams.three;
+package mb.pie.lang.test.call.pieFunc.three;
 
 import mb.pie.api.ExecException;
-import mb.pie.lang.test.call.contextParams.three.DaggerthreeComponent;
+import mb.pie.lang.test.call.pieFunc.three.DaggerthreeComponent;
 import org.junit.jupiter.api.Test;
 
 import static mb.pie.lang.test.util.SimpleChecker.assertTaskOutputEquals;
 
 class ThreeTest {
     @Test void test() throws ExecException {
-        assertTaskOutputEquals(DaggerthreeComponent.class, 4);
+        assertTaskOutputEquals(DaggerthreeComponent.class, 3);
     }
 }
