+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/three/threeComponent.java
@@ -1,12 +1,12 @@
-package mb.pie.lang.test.call.contextParams.one;
+package mb.pie.lang.test.call.pieFunc.three;
 
 import dagger.Component;
 import mb.pie.dagger.PieComponent;
 import mb.pie.dagger.PieModule;
-import mb.pie.lang.test.call.contextParams.one.main_one;
+import mb.pie.lang.test.call.pieFunc.three.main_three;
 
 @mb.pie.dagger.PieScope
 @Component(modules = {PieModule.class, PieTestModule.class}, dependencies = {mb.log.dagger.LoggerComponent.class, mb.resource.dagger.ResourceServiceComponent.class})
-public interface oneComponent extends PieComponent {
-    main_one get();
+public interface threeComponent extends PieComponent {
+    main_three get();
 }
