+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/two/PieTestModule.java
@@ -1,12 +1,12 @@
-package mb.pie.lang.test.call.contextParams.three;
+package mb.pie.lang.test.call.pieFunc.two;
 
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.ElementsIntoSet;
 import mb.pie.api.TaskDef;
-import mb.pie.lang.test.call.contextParams.empty.helper_empty;
-import mb.pie.lang.test.call.contextParams.empty.main_empty;
-import mb.pie.lang.test.call.contextParams.three.main_three;
+import mb.pie.lang.test.call.pieFunc.empty.helper_empty;
+import mb.pie.lang.test.call.pieFunc.empty.main_empty;
+import mb.pie.lang.test.call.pieFunc.two.main_two;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -15,8 +15,8 @@ import java.util.Set;
 abstract class PieTestModule {
     @Provides @mb.pie.dagger.PieScope @ElementsIntoSet
     public static Set<TaskDef<?, ?>> provideTaskDefs(
-        main_three main,
-        helper_three helper
+        main_two main,
+        helper_two helper
     ) {
         final HashSet<TaskDef<?, ?>> taskDefs = new HashSet<>(2, 1);
         taskDefs.add(main);
