+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/two/TwoTest.java
@@ -1,13 +1,13 @@
-package mb.pie.lang.test.call.contextParams.two;
+package mb.pie.lang.test.call.pieFunc.two;
 
 import mb.pie.api.ExecException;
-import mb.pie.lang.test.call.contextParams.two.DaggertwoComponent;
+import mb.pie.lang.test.call.pieFunc.two.DaggertwoComponent;
 import org.junit.jupiter.api.Test;
 
 import static mb.pie.lang.test.util.SimpleChecker.assertTaskOutputEquals;
 
 class TwoTest {
     @Test void test() throws ExecException {
-        assertTaskOutputEquals(DaggertwoComponent.class, 3);
+        assertTaskOutputEquals(DaggertwoComponent.class, 2);
     }
 }
