+++ lang/lang.test/src/test/java/mb/pie/lang/test/call/pieFunc/two/twoComponent.java
@@ -1,12 +1,12 @@
-package mb.pie.lang.test.call.contextParams.omitted;
+package mb.pie.lang.test.call.pieFunc.two;
 
 import dagger.Component;
 import mb.pie.dagger.PieComponent;
 import mb.pie.dagger.PieModule;
-import mb.pie.lang.test.call.contextParams.omitted.main_omitted;
+import mb.pie.lang.test.call.pieFunc.two.main_two;
 
 @mb.pie.dagger.PieScope
 @Component(modules = {PieModule.class, PieTestModule.class}, dependencies = {mb.log.dagger.LoggerComponent.class, mb.resource.dagger.ResourceServiceComponent.class})
-public interface omittedComponent extends PieComponent {
-    main_omitted get();
+public interface twoComponent extends PieComponent {
+    main_two get();
 }
